+++ DVLD_Tests/Applications/ApplicationType/UpdateApplicationTypeServiceTest.cs
@@ -74,7 +74,7 @@ public class UpdateApplicationTypeServiceTest
     [Theory]
     [InlineData(-1)]
     [InlineData(0)]
-    public async Task UpdateAsync_InvalidId_ThrowsArgumentException(int id)
+    public async Task UpdateAsync_InvalidId_ThrowsArgumentException(byte id)
     {
         //Arrange
         ApplicationTypeDTO applicationDTO = new()
